trying to fix shape collabsible (fail)
i'm trying to make the collapsible faq menu such that when it's collapsed, the panel title has rounded corners at 4 sides, but when it's opened that the lower corners are straight (to make it fit more nicely with the text box below).

trying to copy what they're doing here: https://www.digitalocean.com/community/tutorials/css-collapsible  but with no success so far.